+++ sheets/management/commands/process_coupon_assignment_sheet.py
@@ -5,7 +5,7 @@ based on the sheet data, and sends a message to all recipients who received a co
 
 from django.core.management import BaseCommand, CommandError
 
-from sheets.management.utils import assign_coupons_from_spreadsheet
+from sheets.utils import assign_coupons_from_spreadsheet
 from sheets.exceptions import CouponAssignmentError
