+++ static/scss/ecommerce-admin.scss
@@ -162,3 +162,12 @@ button.btn.btn-gradient-white-to-blue:hover {
     max-width: 600px;
   }
 }
+
+.checkbox-div {
+  padding: 0 !important;
+
+  label {
+    padding-left: 5px;
+    margin: 0;
+  }
+}
